+++ tests/settings/test_ce_calculator.py
@@ -162,8 +162,8 @@ def get_spacegroup(db_name):
         concentration=concentration,
         db_name=db_name,
         max_cluster_dia=[4.0, 4.0],
+        include_background_atoms=True,
     )
-    settings.include_background_atoms = True
 
     atoms = crystal(
         symbols=["O", "X", "O", "Ta"],